+++ wazo_calld/plugins/calls/plugin.py
@@ -17,6 +17,7 @@ from .http import (
     CallDtmfResource,
     CallHoldResource,
     CallUnholdResource,
+    CallRecordStartResource,
     CallRecordStopResource,
     CallsResource,
     CallAnswerResource,
@@ -28,6 +29,7 @@ from .http import (
     MyCallDtmfResource,
     MyCallHoldResource,
     MyCallUnholdResource,
+    MyCallRecordStartResource,
     MyCallRecordStopResource,
     MyCallAnswerResource,
 )
@@ -87,6 +89,7 @@ class Plugin:
         api.add_resource(CallDtmfResource, '/calls/<call_id>/dtmf', resource_class_args=[calls_service])
         api.add_resource(CallHoldResource, '/calls/<call_id>/hold/start', resource_class_args=[calls_service])
         api.add_resource(CallUnholdResource, '/calls/<call_id>/hold/stop', resource_class_args=[calls_service])
+        api.add_resource(CallRecordStartResource, '/calls/<call_id>/record/start', resource_class_args=[calls_service])
         api.add_resource(CallRecordStopResource, '/calls/<call_id>/record/stop', resource_class_args=[calls_service])
         api.add_resource(CallAnswerResource, '/calls/<call_id>/answer', resource_class_args=[calls_service])
         api.add_resource(MyCallResource, '/users/me/calls/<call_id>', resource_class_args=[auth_client, calls_service])
@@ -95,6 +98,7 @@ class Plugin:
         api.add_resource(MyCallDtmfResource, '/users/me/calls/<call_id>/dtmf', resource_class_args=[auth_client, calls_service])
         api.add_resource(MyCallHoldResource, '/users/me/calls/<call_id>/hold/start', resource_class_args=[auth_client, calls_service])
         api.add_resource(MyCallUnholdResource, '/users/me/calls/<call_id>/hold/stop', resource_class_args=[auth_client, calls_service])
+        api.add_resource(MyCallRecordStartResource, '/users/me/calls/<call_id>/record/start', resource_class_args=[auth_client, calls_service])
         api.add_resource(MyCallRecordStopResource, '/users/me/calls/<call_id>/record/stop', resource_class_args=[auth_client, calls_service])
         api.add_resource(MyCallAnswerResource, '/users/me/calls/<call_id>/answer', resource_class_args=[auth_client, calls_service])
         api.add_resource(ConnectCallToUserResource, '/calls/<call_id>/user/<user_uuid>', resource_class_args=[calls_service])