+++ wazo_calld/plugins/calls/services.py
@@ -359,6 +359,23 @@ class CallsService:
         self._verify_user(call_id, user_uuid)
         self.unhold(call_id)
 
+    def record_start(self, call_id):
+        try:
+            channel = self._ari.channels.get(channelId=call_id)
+        except ARINotFound:
+            raise NoSuchCall(call_id)
+
+        ami.record_start(self._ami, call_id)
+
+        # NOTE(afournier): asterisk should send back an event instead of
+        # wrongly pretend that the channel has started recording
+        call = self.make_call_from_channel(self._ari, channel)
+        self._notifier.call_updated(call)
+
+    def record_start_user(self, call_id, user_uuid):
+        self._verify_user(call_id, user_uuid)
+        self.record_start(call_id)
+
     def record_stop(self, call_id):
         try:
             channel = self._ari.channels.get(channelId=call_id)